(/) static cast added to solve compil error

+++ client/src/Game/Game.cpp
@@ -309,7 +309,7 @@ void Game::update()
     if (std::chrono::system_clock::now() - this->_lastPing > std::chrono::seconds(1)) {
         this->_lastPing = std::chrono::system_clock::now();
         Stream out;
-        out << 12_uc << std::chrono::duration_cast<std::chrono::milliseconds>(_lastPing.time_since_epoch()).count();
+        out << 12_uc << static_cast<long>(std::chrono::duration_cast<std::chrono::milliseconds>(_lastPing.time_since_epoch()).count());
         this->_net.send(out);
     }
